moved Mapper files to application package. refactored UserMapper
